+++ projects/ngx-scroll-animations/src/lib/ngx-scroll-animations.directive.ts
@@ -9,24 +9,28 @@ import {
   NgZone,
   OnDestroy,
   OnInit,
+  Optional,
   Output,
   PLATFORM_ID,
   Renderer2,
 } from '@angular/core';
-import { Observable, Subject, fromEvent, of } from 'rxjs';
 import {
+  Observable,
+  Subject,
   debounceTime,
-  delay,
   distinctUntilChanged,
   filter,
-  first,
+  fromEvent,
   map,
+  of,
   startWith,
   switchMap,
   takeUntil,
   takeWhile,
   tap,
-} from 'rxjs/operators';
+  timer,
+} from 'rxjs';
+import { NGX_SCROLL_ANIMATION_CONFIGS } from './ngx-scroll-animations.module';
 import { NgxScrollAnimationsService } from './ngx-scroll-animations.service';
 import {
   BooleanInput,
@@ -36,9 +40,11 @@ import {
   NumberInput,
   coerceNumberProperty,
 } from './utils/coercion/coercion-number';
+import { NgxScrollAnimationConfigs } from './utils/ngx-scroll-animation-configs';
 import { ThresholdModeT } from './utils/ngx-scroll-animations-types';
 
 @Directive({
+  standalone: true,
   selector: '[ngxScrollAnimate]',
 })
 export class NgxScrollAnimationsDirective
@@ -55,16 +61,13 @@ export class NgxScrollAnimationsDirective
   @Output() endAnimation = new EventEmitter<void>();
 
   /**
-   * Defines the available speed options for the animation.
-   */
-
-  /**
    * The speed at which the animation runs in milliseconds.
    * @param speed - The desired speed of the animation, chosen from predefined options.
+   * @default 300
    */
   @Input()
-  public set speed(speed: number | string | undefined) {
-    this._speed = `${+(speed ?? 300)}ms`;
+  public set speed(speed: NumberInput) {
+    this._speed = `${coerceNumberProperty(speed, 300)}ms`;
   }
   public get speed(): string {
     return this._speed;
@@ -74,22 +77,18 @@ export class NgxScrollAnimationsDirective
   /**
    * Delays the start of the animation. Accepts the delay time in milliseconds.
    * @param delayTime - The time in milliseconds to delay the start of the animation.
+   * @default 0
    */
   @Input()
   set delay(delayTime: NumberInput) {
     const value = coerceNumberProperty(delayTime);
-    if (value) {
-      this._delay = `${value}ms`;
-    } else {
-      this._delay = delayTime ? `${delayTime}ms` : '';
-    }
+    this._delay = !!value ? `${value}ms` : delayTime?.toString() ?? '';
   }
   public get delay(): string {
     return this._delay;
   }
   private _delay: string = '';
 
-  private _disabled: boolean = false;
   /**
    * A boolean value to enable or disable the animation.
    */
@@ -100,29 +99,38 @@ export class NgxScrollAnimationsDirective
   get disabled(): boolean {
     return this._disabled;
   }
+  private _disabled: boolean = false;
 
   /**
    * Defines how the animation accelerates and decelerates during its runtime.
-   * Acceptable values: 'ease', 'ease-in', 'ease-out', 'ease-in-out', or a 'cubic-bezier()' function call.
+   *
+   * Example values: `ease`, `ease-in`, `ease-out`, `ease-in-out`, `cubic-bezier()`
+   *
+   * @default `ease`
    */
   @Input()
   easing: 'ease' | 'ease-in' | 'ease-out' | 'ease-in-out' | string = 'ease';
 
   /**
    * The threshold for triggering the animation when an element scrolls into the viewport.
-   * The default is 80% for 'percent' mode and 20 for 'pixel' mode if not specified.
+   *
+   * The default is `80%` for `percent` mode and `20` for `pixel` mode if not specified.
    */
   @Input()
-  public set aos(value: number | string | undefined) {
-    this.threshold = value
-      ? +value
-      : this.thresholdMode === 'percent'
-        ? 0.8
-        : 20;
+  public set aos(value: NumberInput) {
+    if (!!value && typeof value !== 'number' && isNaN(+value)) {
+      throw new Error('`aos` must be a number');
+    }
+    this.threshold = coerceNumberProperty(
+      value,
+      this.thresholdMode === 'percent' ? 0.8 : 20
+    );
   }
 
   /**
-   * Defines the mode for calculating the threshold: 'percent' or 'pixel'.
+   * Defines the mode for calculating the threshold: `percent` or `pixel`.
+   *
+   * @default `percent`
    */
   @Input()
   set thresholdMode(mode: ThresholdModeT | undefined) {
@@ -133,9 +141,10 @@ export class NgxScrollAnimationsDirective
   }
   private _thresholdMode: ThresholdModeT = 'percent';
 
-  private _once: boolean = true;
   /**
-   * If true, triggers the animation only once when the element scrolls into the viewport.
+   * If `true`, triggers the animation only once when the element scrolls into the viewport.
+   *
+   * @default true
    */
   @Input()
   set once(value: BooleanInput) {
@@ -144,10 +153,12 @@ export class NgxScrollAnimationsDirective
   get once(): boolean {
     return this._once;
   }
+  private _once: boolean = true;
 
-  private _zoneless: boolean = true;
   /**
-   * Set this property to false if your application runs with ng zone.
+   * Set this property to `false` if your application runs with ng zone.
+   *
+   * @default true
    */
   @Input()
   set zoneless(value: BooleanInput) {
@@ -156,10 +167,13 @@ export class NgxScrollAnimationsDirective
   get zoneless(): boolean {
     return this._zoneless;
   }
+  private _zoneless: boolean = true;
 
-  private _undoGap: number = 20;
   /**
-   * The gap between the animation start point and animation leave point.
+   * The gap between the animation start point and the animation end point.
+   * This prevents the animation from starting/stopping improperly,
+   * which often happens on mobile devices.
+   *
    * @default 20
    */
   @Input()
@@ -169,47 +183,36 @@ export class NgxScrollAnimationsDirective
   get undoGap(): number {
     return this._undoGap;
   }
-
-  /**
-   * When set to true, replays the animation. Useful for re-triggering animations.
-   */
-  @Input()
-  public set replay(replay: BooleanInput) {
-    // Re-triggers the animation again on request (skipping the very fist value)
-    if (coerceBooleanProperty(replay)) {
-      this.triggerIdle();
-      this.replay$.next(true);
-    }
-  }
+  private _undoGap: number = 20;
 
   /**
    * Specifies the animation to be played.
+   *
+   * @default
+   * 'fade-in-up'
    */
   @Input('ngxScrollAnimate')
-  set animate(val: string | undefined) {
-    if (val) {
-      this._animate = val;
-    }
+  set animationName(val: string | undefined) {
+    if (val) this._animate = val;
   }
-  public get animate(): string {
+  public get animationName(): string {
     return this._animate;
   }
   private _animate: string = 'fade-in-up';
 
-  /** Subject to manage replaying of the animation */
-  private replay$ = new Subject<boolean>();
-
-  /** Flag to indicate if the animation is currently running */
+  /** Indicates how many time the animation is active triggered on the current element */
   public activeAnimations = 0;
+
+  /** Total count how many time the animation is triggered */
+  public count = 0;
+
+  public lastViewState?: VisibleState;
 
   /** Threshold of the animation */
   private threshold = 0.8;
 
-  private count = 0;
-  private lastViewState?: number;
-
-  private onStable = new EventEmitter<any>();
   private temporaryBoundings!: DOMRect;
+  private onStable = new EventEmitter<any>();
   private destroy$ = new Subject();
 
   constructor(
@@ -217,34 +220,82 @@ export class NgxScrollAnimationsDirective
     private zone: NgZone,
     private elRef: ElementRef<HTMLElement>,
     private scrollService: NgxScrollAnimationsService,
-    @Inject(PLATFORM_ID) private platformId: any,
-  ) {}
+    @Optional()
+    @Inject(NGX_SCROLL_ANIMATION_CONFIGS)
+    configs: NgxScrollAnimationConfigs | undefined,
+    @Inject(PLATFORM_ID) private platformId: any
+  ) {
+    if (configs) {
+      if (configs.speed !== undefined) {
+        this.speed = configs.speed;
+      }
+      if (configs.delay !== undefined) {
+        this.delay = configs.delay;
+      }
+      if (configs.disabled !== undefined) {
+        this.disabled = configs.disabled;
+      }
+      if (configs.easing !== undefined) {
+        this.easing = configs.easing;
+      }
+      if (configs.threshold !== undefined) {
+        this.aos = configs.threshold;
+      }
+      if (configs.thresholdMode !== undefined) {
+        this.thresholdMode = configs.thresholdMode;
+      }
+      if (configs.once !== undefined) {
+        this.once = configs.once;
+      }
+      if (configs.zoneless !== undefined) {
+        this.zoneless = configs.zoneless;
+      }
+      if (configs.undoGap !== undefined) {
+        this.undoGap = configs.undoGap;
+      }
+      if (configs.animationName !== undefined) {
+        this.animationName = configs.animationName;
+      }
+    }
+  }
 
   /**
    * Initializes the directive and sets up the animation triggers.
    */
   ngOnInit(): void {
-    if (!this.zoneless) this.onStable = this.zone.onStable;
-
     this.setTemporaryBoundings();
     this.elRef.nativeElement.classList.add('ngx-scroll-animations');
-    this.zone.runOutsideAngular(() => {
-      this.triggerIdle();
-      if (isPlatformBrowser(this.platformId)) {
-        this.listenAnimationState();
-        this.setupAnimationTrigger();
-      }
-    });
+
+    if (this.zoneless) {
+      this.initListeners();
+    } else {
+      this.onStable = this.zone.onStable;
+      this.zone.runOutsideAngular(() => {
+        this.initListeners();
+      });
+    }
   }
 
   ngAfterViewInit(): void {
-    if (this.zoneless) this.onStable.next(true);
+    if (this.zoneless) {
+      timer(1).subscribe(() => {
+        this.onStable.next(true);
+      });
+    }
     this.setTemporaryBoundings();
   }
 
   ngOnDestroy(): void {
     this.destroy$.next(true);
     this.destroy$.complete();
+  }
+
+  private initListeners() {
+    this.triggerIdle();
+    if (isPlatformBrowser(this.platformId)) {
+      this.listenAnimationState();
+      this.setupAnimationTrigger();
+    }
   }
 
   private setTemporaryBoundings() {
@@ -260,9 +311,9 @@ export class NgxScrollAnimationsDirective
   private listenAnimationState(): void {
     fromEvent(this.elRef.nativeElement, 'animationstart')
       .pipe(
-        takeUntil(this.destroy$),
         // Make sure the animation is not triggered by a child element.
         filter((event) => event.target === this.elRef.nativeElement),
+        takeUntil(this.destroy$)
       )
       .subscribe(() => {
         this.startAnimation.emit();
@@ -270,14 +321,14 @@ export class NgxScrollAnimationsDirective
 
     fromEvent(this.elRef.nativeElement, 'animationend')
       .pipe(
-        takeUntil(this.destroy$),
         // Make sure the animation is not triggered by a child element.
         filter((event) => event.target === this.elRef.nativeElement),
         tap(() => {
           this.endAnimation.emit();
-          this.clearAnimation(this.once || this.lastViewState === 1);
+          this.clearAnimation(this.once || this.lastViewState === 'visible');
         }),
         debounceTime(10),
+        takeUntil(this.destroy$)
       )
       .subscribe(() => {
         this.activeAnimations--;
@@ -288,47 +339,34 @@ export class NgxScrollAnimationsDirective
    * Configures the RxJS stream for triggering animations.
    */
   private setupAnimationTrigger(): void {
-    this.replay$
+    this.onStable
       .pipe(
-        () =>
-          this.onStable.pipe(
-            first(),
-            delay(1), // Delay is important in zoneless applications
-            switchMap((trigger) =>
-              this.threshold > 0
-                ? this.scrollService.scroll$!.pipe(
-                    startWith(0),
-                    switchMap(() =>
-                      this.checkVisibility(
-                        this.elRef.nativeElement,
-                        this.threshold,
-                        this.thresholdMode,
-                        this.undoGap,
-                      ),
-                    ),
-                  )
-                : of(trigger),
-            ),
-          ),
+        // Check element is visible
+        switchMap(() => this.checkVisibility()),
+        // Ignore same states
         distinctUntilChanged(),
-        filter(
-          (val) =>
-            val !== undefined &&
-            (this.activeAnimations === 0 || this.lastViewState === 0) &&
-            (this.lastViewState !== val ||
-              (this.once && this.lastViewState !== 1)),
-        ),
-        takeWhile((trigger) => !trigger || !this.once, true),
-        takeUntil(this.destroy$),
+        // Do nothing when the position is `hold`
+        filter((val) => val !== 'hold'),
+        // If the animation run only once we can stop after the element is visible
+        takeWhile((trigger) => trigger !== 'visible' || !this.once, true),
+        // Update the lastViewState
+        tap((trigger) => (this.lastViewState = trigger)),
+        // Check in which direction the animation should run
+        map((trigger) => {
+          if (trigger === 'visible') {
+            return 'normal';
+          } else if (this.count > 0) {
+            return 'reverse';
+          }
+          return undefined;
+        }),
+        // Destroy the animation listener
+        takeUntil(this.destroy$)
       )
       .subscribe({
-        next: (trigger) => {
-          this.lastViewState = trigger;
-
-          if (trigger) {
-            this.triggerAnimation('normal');
-          } else if (!this.once && this.count > 0) {
-            this.triggerAnimation('reverse');
+        next: (direction) => {
+          if (direction) {
+            this.triggerAnimation(direction);
           } else {
             this.triggerIdle();
           }
@@ -355,9 +393,9 @@ export class NgxScrollAnimationsDirective
       this.renderer.setStyle(
         this.elRef.nativeElement,
         'animation',
-        `${this.animate} ${this.speed} ${this.delay ? this.delay + ' ' : ''}${
-          this.easing
-        } ${direction}`,
+        `${this.animationName} ${this.speed} ${
+          this.delay ? this.delay + ' ' : ''
+        }${this.easing} ${direction}`
       );
     }
   }
@@ -393,28 +431,40 @@ export class NgxScrollAnimationsDirective
    * @param undoGap - The gap between the animation start point and animation leave point.
    * @returns An Observable emitting the visibility state (0, 1, or undefined).
    */
-  private checkVisibility(
-    element: HTMLElement,
-    threshold: number,
-    thresholdMode: ThresholdModeT,
-    undoGap: number,
-  ): Observable<0 | 1 | undefined> {
-    return this.scrollService.view$!.pipe(
-      map((view) => {
-        let rect!: DOMRect;
-        if (this.activeAnimations > 0) {
-          rect = this.temporaryBoundings;
-        } else {
-          rect = element.getBoundingClientRect();
-          this.temporaryBoundings = rect;
-        }
+  private checkVisibility(): Observable<VisibleState> {
+    if (this.threshold <= 0) return of('visible');
 
-        const triggerPos =
-          thresholdMode === 'percent' ? rect.height * threshold : threshold;
+    return this.scrollService.scroll$.pipe(
+      startWith(0),
+      switchMap(() =>
+        this.scrollService.view$!.pipe(
+          map((view) => {
+            let rect!: DOMRect;
+            if (this.activeAnimations > 0) {
+              rect = this.temporaryBoundings;
+            } else {
+              rect = this.elRef.nativeElement.getBoundingClientRect();
+              this.temporaryBoundings = rect;
+            }
 
-        const currentPos = rect.top - view.height + triggerPos;
-        return currentPos <= 0 ? 1 : currentPos > undoGap ? 0 : undefined;
-      }),
+            const triggerPos =
+              this.thresholdMode === 'percent'
+                ? rect.height * this.threshold
+                : this.threshold;
+
+            const currentPos = rect.top - view.height + triggerPos;
+            if (currentPos <= 0) {
+              return 'visible';
+            } else if (currentPos > this.undoGap) {
+              return 'hidden';
+            } else {
+              return 'hold';
+            }
+          })
+        )
+      )
     );
   }
 }
+
+type VisibleState = 'visible' | 'hidden' | 'hold';