+++ projects/ngx-scroll-animations/src/lib/ngx-scroll-animations.module.ts
@@ -1,12 +1,37 @@
-import { NgModule } from '@angular/core';
+import {
+  InjectionToken,
+  ModuleWithProviders,
+  NgModule,
+  Provider,
+} from '@angular/core';
 import { NgxScrollAnimationsDirective } from './ngx-scroll-animations.directive';
-import { NgxScrollAnimationsService } from './ngx-scroll-animations.service';
+import { NgxScrollAnimationConfigs } from './utils/ngx-scroll-animation-configs';
 
+export const NGX_SCROLL_ANIMATION_CONFIGS =
+  new InjectionToken<NgxScrollAnimationConfigs>('NgxScrollAnimationConfigs');
 
 @NgModule({
-    imports: [],
-    exports: [NgxScrollAnimationsDirective],
-    declarations: [NgxScrollAnimationsDirective],
-    providers: [NgxScrollAnimationsService],
+  imports: [NgxScrollAnimationsDirective],
+  exports: [NgxScrollAnimationsDirective],
 })
-export class NgxScrollAnimationsModule { }
+export class NgxScrollAnimationsModule {
+  static forRoot(
+    configs: NgxScrollAnimationConfigs
+  ): ModuleWithProviders<NgxScrollAnimationsModule> {
+    return {
+      ngModule: NgxScrollAnimationsModule,
+      providers: [provideNgxScrollAnimations(configs)],
+    };
+  }
+}
+
+export function provideNgxScrollAnimations(
+  configs?: NgxScrollAnimationConfigs
+): Provider {
+  return [
+    {
+      provide: NGX_SCROLL_ANIMATION_CONFIGS,
+      useValue: configs,
+    },
+  ];
+}