+++ projects/ngx-scroll-animations/src/lib/ngx-scroll-animations.service.ts
@@ -1,14 +1,15 @@
 import { DOCUMENT, isPlatformBrowser } from '@angular/common';
 import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
-import { Observable, fromEvent } from 'rxjs';
 import {
+  Observable,
   debounceTime,
+  fromEvent,
   map,
   share,
   shareReplay,
   startWith,
-  throttleTime
-} from 'rxjs/operators';
+  throttleTime,
+} from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
@@ -19,21 +20,22 @@ export class NgxScrollAnimationsService {
 
   constructor(
     @Inject(DOCUMENT) private document: Document,
-    @Inject(PLATFORM_ID) private platformId: any,
+    @Inject(PLATFORM_ID) private platformId: any
   ) {
     if (isPlatformBrowser(this.platformId)) {
       // Observable for viewport changes, using window resize events as a trigger
       this.view$ = fromEvent(window, 'resize').pipe(
         debounceTime(100),
         startWith(this.getCurrentViewport()),
-        map(() => this.getCurrentViewport()),
-        shareReplay(1),
+        map(this.getCurrentViewport),
+        // Share the last screen size
+        shareReplay(1)
       );
 
       // Observable for scroll events on the window
       this.scroll$ = fromEvent(this.document, 'scroll').pipe(
         throttleTime(50, undefined, { leading: true, trailing: true }),
-        share(),
+        share()
       );
     }
   }
@@ -47,7 +49,22 @@ export class NgxScrollAnimationsService {
   public overrideScrollListener(element: HTMLElement): void {
     this.scroll$ = fromEvent(element, 'scroll').pipe(
       throttleTime(50, undefined, { leading: true, trailing: true }),
-      share(),
+      share()
+    );
+  }
+  /**
+   * Overrides the default resize listener for the specified HTML element.
+   *
+   * @param {HTMLElement} element - The HTML element for which to override the resize listener.
+   * @returns {void}
+   */
+  public overrideResizeListener(element: HTMLElement): void {
+    this.view$ = fromEvent(element, 'resize').pipe(
+      debounceTime(100),
+      startWith(this.getCurrentViewport()),
+      map(this.getCurrentViewport),
+      // Share the last screen size
+      shareReplay(1)
     );
   }