+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+
+## `2.0.0` - 2024.06.27
+### Feat
+- Convert to a complete standalone solution for better support with Angular 18.
+- Added ability to set global configs.
+
 ## `1.2.2` - 2024.06.20
 ### Fix
 - Resolved issue where scrolling did not retrigger when a child element started an animation.